+++ packages/editor/editor-core/src/__tests__/visual-regression/_utils.ts
@@ -172,7 +172,9 @@ export const baseTests = [
   },
   {
     name: 'Action',
-    clickSelector: `span[aria-label="${listsMessages.action.defaultMessage}"]`,
+    clickSelector: `span[aria-label="${
+      insertBlockMessages.action.defaultMessage
+    }"]`,
     nodeSelector: 'ol[data-task-list-local-id] div',
     content: 'text',
     appearance: ['full-page', 'message'],