+++ packages/editor/editor-core/src/plugins/table/ui/TableFloatingControls/CornerControls/index.tsx
@@ -10,12 +10,14 @@ import { TableCssClassName as ClassName } from '../../../types';
 export interface Props {
   editorView: EditorView;
   selection?: Selection;
+  tableRef?: HTMLElement;
   clearHoverSelection: () => void;
   isTableInDanger?: boolean;
   isHeaderColumnEnabled?: boolean;
   isHeaderRowEnabled?: boolean;
   isNumberColumnEnabled?: boolean;
-  showInsertButton?: boolean;
+  insertColumnButtonIndex?: number;
+  insertRowButtonIndex?: number;
 }
 
 export default class CornerControls extends Component<Props, any> {
@@ -29,8 +31,9 @@ export default class CornerControls extends Component<Props, any> {
       isTableInDanger,
       isHeaderRowEnabled,
       isHeaderColumnEnabled,
-      isNumberColumnEnabled,
-      showInsertButton,
+      insertColumnButtonIndex,
+      insertRowButtonIndex,
+      tableRef,
     } = this.props;
 
     return (
@@ -48,20 +51,21 @@ export default class CornerControls extends Component<Props, any> {
           onMouseOver={this.hoverTable}
           onMouseOut={this.props.clearHoverSelection}
         />
-        {!isHeaderColumnEnabled &&
-          !isNumberColumnEnabled && (
-            <InsertButton
-              type="column"
-              index={0}
-              showInsertButton={showInsertButton}
-              onMouseDown={this.insertColumn}
-            />
-          )}
+        {!isHeaderColumnEnabled && (
+          <InsertButton
+            type="column"
+            tableRef={tableRef!}
+            index={0}
+            showInsertButton={insertColumnButtonIndex === 0}
+            onMouseDown={this.insertColumn}
+          />
+        )}
         {!isHeaderRowEnabled && (
           <InsertButton
             type="row"
+            tableRef={tableRef!}
             index={0}
-            showInsertButton={showInsertButton}
+            showInsertButton={insertRowButtonIndex === 0}
             onMouseDown={this.insertRow}
           />
         )}