+++ packages/elements/analytics-namespaced-context/src/__tests__/NavigationAnalyticsContextSpec.tsx
@@ -47,7 +47,6 @@ describe('<NavigationAnalyticsContext />', () => {
             packageVersion: '1.0.0',
           },
           eventType: 'ui',
-          source: 'unknown',
         },
       }),
       FabricChannel.navigation,