+++ packages/media/media-card/src/root/card/cardLoader.tsx
@@ -7,7 +7,8 @@ interface AsyncCardProps {
 }
 
 export default class Card extends React.PureComponent<
-  CardProps & AsyncCardProps
+  CardProps & AsyncCardProps,
+  AsyncCardProps
 > {
   static displayName = 'AsyncCard';
   static Card?: typeof CardType;
@@ -27,9 +28,14 @@ export default class Card extends React.PureComponent<
   }
 
   render() {
+    const { dimensions, identifier } = this.props;
+
     if (!this.state.Card) {
       return (
-        <CardLoading mediaItemType={this.props.identifier.mediaItemType} />
+        <CardLoading
+          dimensions={dimensions}
+          mediaItemType={identifier.mediaItemType}
+        />
       );
     }